organize entities into folders
